+++ iocore/hostdb/Makefile.am
@@ -40,7 +40,10 @@ libinkhostdb_a_SOURCES = \
 	P_HostDBProcessor.h \
 	P_RefCountCache.h \
 	P_RefCountCacheSerializer.h \
-	RefCountCache.cc
+	RefCountCache.cc \
+	HostFile.h \
+	HostFile.cc \
+	HostDBInfo.cc
 
 TESTS = $(check_PROGRAMS)
 check_PROGRAMS = test_RefCountCache test_HostFile
@@ -88,39 +91,44 @@ test_HostFile_CPPFLAGS = \
 	-I$(abs_top_srcdir)/proxy/http/remap
 
 test_HostFile_LDADD = \
-	$(top_builddir)/proxy/http/libhttp.a \
-	$(top_builddir)/proxy/http/remap/libhttp_remap.a \
-	$(top_builddir)/proxy/http2/libhttp2.a \
-	$(top_builddir)/proxy/logging/liblogging.a \
-	$(top_builddir)/proxy/hdrs/libhdrs.a \
-	$(top_builddir)/proxy/shared/libdiagsconfig.a \
-	$(top_builddir)/mgmt/libmgmt_p.la \
-	$(top_builddir)/iocore/utils/libinkutils.a \
-	$(top_builddir)/iocore/dns/libinkdns.a \
-	$(top_builddir)/iocore/hostdb/libinkhostdb.a \
-	$(top_builddir)/iocore/dns/libinkdns.a \
-	$(top_builddir)/iocore/cache/libinkcache.a \
-	$(top_builddir)/lib/fastlz/libfastlz.a \
-	$(top_builddir)/iocore/aio/libinkaio.a \
-	$(top_builddir)/src/tscore/libtscore.la \
-	$(top_builddir)/src/tscpp/util/libtscpputil.la \
-	$(top_builddir)/proxy/libproxy.a \
-	$(top_builddir)/iocore/net/libinknet.a \
-	$(top_builddir)/lib/records/librecords_p.a \
-	$(top_builddir)/iocore/eventsystem/libinkevent.a \
-	@HWLOC_LIBS@ \
-	@LIBPCRE@ \
-	@LIBRESOLV@ \
-	@LIBZ@ \
-	@LIBLZMA@ \
-	@LIBPROFILER@ \
-	@OPENSSL_LIBS@ \
-	@YAMLCPP_LIBS@ \
-	@BORINGOCSP_LIBS@ \
-	-lm
+	$(top_builddir)/src/tscore/libtscore.la
+
+# $(top_builddir)/proxy/http/libhttp.a \
+# $(top_builddir)/proxy/http/remap/libhttp_remap.a \
+# $(top_builddir)/proxy/http2/libhttp2.a \
+# $(top_builddir)/proxy/logging/liblogging.a \
+# $(top_builddir)/proxy/hdrs/libhdrs.a \
+# $(top_builddir)/proxy/shared/libdiagsconfig.a \
+# $(top_builddir)/mgmt/libmgmt_p.la \
+# $(top_builddir)/iocore/utils/libinkutils.a \
+# $(top_builddir)/iocore/dns/libinkdns.a \
+# $(top_builddir)/iocore/hostdb/libinkhostdb.a \
+# $(top_builddir)/iocore/dns/libinkdns.a \
+# $(top_builddir)/iocore/cache/libinkcache.a \
+# $(top_builddir)/lib/fastlz/libfastlz.a \
+# $(top_builddir)/iocore/aio/libinkaio.a \
+# $(top_builddir)/src/tscore/libtscore.la \
+# $(top_builddir)/src/tscpp/util/libtscpputil.la \
+# $(top_builddir)/proxy/libproxy.a \
+# $(top_builddir)/iocore/net/libinknet.a \
+# $(top_builddir)/lib/records/librecords_p.a \
+# $(top_builddir)/iocore/eventsystem/libinkevent.a \
+# @HWLOC_LIBS@ \
+# @LIBPCRE@ \
+# @LIBRESOLV@ \
+# @LIBZ@ \
+# @LIBLZMA@ \
+# @LIBPROFILER@ \
+# @OPENSSL_LIBS@ \
+# @YAMLCPP_LIBS@ \
+# @BORINGOCSP_LIBS@ \
+# -lm
 
 test_HostFile_SOURCES = \
-	test_HostFile.cc
+	test_HostFile.cc \
+	HostFile.h \
+	HostFile.cc \
+	HostDBInfo.cc
 
 include $(top_srcdir)/build/tidy.mk
